database

+++ lib/constants.dart
@@ -3,10 +3,10 @@ const int kUpgrade2Price = 1000;
 const int kUpgrade3Price = 5000;
 const int kUpgrade4Price = 25000;
 
-const double kUpgrade1Multiplier = 2;
-const double kUpgrade2Multiplier = 5;
-const double kUpgrade3Multiplier = 10;
-const double kUpgrade4Multiplier = 50;
+const int kUpgrade1Multiplier = 2;
+const int kUpgrade2Multiplier = 5;
+const int kUpgrade3Multiplier = 10;
+const int kUpgrade4Multiplier = 50;
 
 const int kInsanePrice = 1000000;
-const double kInsaneMultiplier = 1000000;
+const int kInsaneMultiplier = 1000000;
